pillow_sample/requirements.txt
arabic-reshaper==2.1.3
future==0.18.2
Pillow==9.1.1
python-bidi==0.4.2
six==1.16.0
